backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
pandas
statsmodels
numpy
scipy
scikit-learn
gunicorn==21.2.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
python-dotenv==1.0.0
